cleanup

+++ Automaton.java
@@ -63,6 +63,7 @@ public class Automaton {
     //number to binary, ternary etc. based on kval.
     public static String[] intToKAry(int code, int kVal) {
         String[] kCode = new String[(3*kVal-2)];
+        // int newCode = code + ((int)Math.pow(kVal,((3*kVal)-2))); //Get to center(ish) of ks
         for(int n = 0; n < kCode.length; n++) {
             String digit = Integer.toString((code/((int)Math.pow(kVal,n)))%kVal);
             kCode[kCode.length - n - 1] = digit;
@@ -95,8 +96,8 @@ public class Automaton {
     // using mapValToColor. This is how color is set generation to
     // generation.
     public static int mapAvgToVal(double avg, int kInt, String[] kCode){
-        double dif = kInt-avg;
-        double index = 3*(dif-1);
+        double dif = (double)kInt-avg;
+        double index = 3.0*(dif-1.0);
         int ind = (int)Math.round(index);
         return Integer.parseInt(kCode[ind]);
     }
